backend/pyproject.toml
[project]
name = "prompter"
version = "0.1.0"
description = "Prompt engineering toolkit"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
]

[project.scripts]
prompter = "prompter.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
